requirements.txt
# core compat for Python 3.13
numpy==2.1.2
typing_extensions>=4.9.0

# openai 1.x needs these pinned on Windows to avoid httpx 'proxies' error
httpx==0.27.2
anyio>=3.5.0,<5
sniffio
tqdm
distro
jiter>=0.4.0

# your libs
pydantic==2.8.2
langchain==0.3.27
langgraph==0.2.34
openai==1.43.0
python-dotenv==1.0.1
boto3==1.34.93